+++ src/framework/piral-core/src/components/ExtensionSlot.tsx
@@ -2,7 +2,11 @@ import * as React from 'react';
 import { isfunc } from 'piral-base';
 import { useGlobalState } from '../hooks';
 import { defaultRender, none } from '../utils';
-import { ExtensionSlotProps } from '../types';
+import { ExtensionRegistration, ExtensionSlotProps } from '../types';
+
+function defaultOrder(extensions: Array<ExtensionRegistration>) {
+  return extensions;
+}
 
 /**
  * The extension slot component to be used when the available
@@ -10,12 +14,12 @@ import { ExtensionSlotProps } from '../types';
  * location.
  */
 export function ExtensionSlot<T extends string>(props: ExtensionSlotProps<T>) {
-  const { name, render = defaultRender, empty, params, children, noEmptyRender = false } = props;
+  const { name, render = defaultRender, empty, params, children, noEmptyRender = false, order = defaultOrder } = props;
   const extensions = useGlobalState((s) => s.registry.extensions[name] || none);
   const isEmpty = extensions.length === 0 && isfunc(empty);
   const content = isEmpty
     ? [defaultRender(empty(), 'empty')]
-    : extensions.map(({ component: Component, reference, defaults = {} }, i) => (
+    : order(extensions).map(({ component: Component, reference, defaults = {} }, i) => (
         <Component
           key={`${reference?.displayName || '_'}${i}`}
           children={children}